+++ changelog.md
@@ -5,6 +5,8 @@
 
 ##v1.2
 -------------------
++ Builder Activity now attempts to update partial course data before building
++ Fixed major bug with scheduler builder when using CRN's
 + Bulk Transactions for Improving Parser Performance
 + Reduced Sleep time for partial parser to improve responsiveness
 + Removed unnecessary transcation calls for data retrieval queries
@@ -14,7 +16,6 @@
 + Added red text coloring to full courses
 + Improved Schedule View Spacing (using weights now)
 + Improved Star View Spacing (using different weights)
-+ Fixed major bug with scheduler builder when using CRN's
 
 ##v1.1 [Currently Published]
 -------------------